requirements.txt
gunicorn
flask==2.3.2
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
protobuf==3.20.*